Move fieldOption template to other template utilities
