--- src/utilities/constants/templates.ts
@@ -1,10 +1,0 @@
-const templates = {
-    NEW_FIELD_OPTION: {
-        local_id: 0,
-        title: '',
-        description: '',
-        mandatory: false,
-    },
-};
-
-export default templates;
